sensor-app/requirements.txt
pyserial
gpiod
paho-mqtt
transitions
